Move get_launch_grouping_type to the service
